[Feature] 즐겨찾기 모달 생성
- 폴짝추가 모달 함께 사용

# 236
